Removed Napatech specific commands:
RTE_FLOW_ITEM_TYPE_GREv0, RTE_FLOW_ITEM_TYPE_GREv1,
RTE_FLOW_ITEM_TYPE_GTPv0_U  and RTE_FLOW_ITEM_TYPE_GTPv1_U
as they are replaced by DPDK commands
RTE_FLOW_ITEM_TYPE_GRE, RTE_FLOW_ITEM_TYPE_GTPC and RTE_FLOW_ITEM_TYPE_GTPU

Command RTE_FLOW_ITEM_TYPE_MPLS supported.

Corrected compile errors on Centos

Napatech example cmbatch optimized

C malloc changed to use DPDK malloc rte_zmalloc in ntacc PMD.

+++ examples/load_balancer/runtime.c
@@ -47,7 +47,6 @@
 #include <rte_log.h>
 #include <rte_memory.h>
 #include <rte_memcpy.h>
-#include <rte_memzone.h>
 #include <rte_eal.h>
 #include <rte_launch.h>
 #include <rte_atomic.h>
@@ -57,7 +56,6 @@
 #include <rte_per_lcore.h>
 #include <rte_branch_prediction.h>
 #include <rte_interrupts.h>
-#include <rte_pci.h>
 #include <rte_random.h>
 #include <rte_debug.h>
 #include <rte_ether.h>
@@ -188,7 +186,7 @@ app_lcore_io_rx(
 	uint32_t i;
 
 	for (i = 0; i < lp->rx.n_nic_queues; i ++) {
-		uint8_t port = lp->rx.nic_queues[i].port;
+		uint16_t port = lp->rx.nic_queues[i].port;
 		uint8_t queue = lp->rx.nic_queues[i].queue;
 		uint32_t n_mbufs, j;
 
@@ -213,7 +211,7 @@ app_lcore_io_rx(
 
 			printf("I/O RX %u in (NIC port %u): NIC drop ratio = %.2f avg burst size = %.2f\n",
 				lcore,
-				(unsigned) port,
+				port,
 				(double) stats.imissed / (double) (stats.imissed + stats.ipackets),
 				((double) lp->rx.nic_queues_count[i]) / ((double) lp->rx.nic_queues_iters[i]));
 			lp->rx.nic_queues_iters[i] = 0;
@@ -339,7 +337,7 @@ app_lcore_io_tx(
 		uint32_t i;
 
 		for (i = 0; i < lp->tx.n_nic_ports; i ++) {
-			uint8_t port = lp->tx.nic_ports[i];
+			uint16_t port = lp->tx.nic_ports[i];
 			struct rte_ring *ring = lp->tx.rings[port][worker];
 			uint32_t n_mbufs, n_pkts;
 			int ret;
@@ -395,7 +393,7 @@ app_lcore_io_tx(
 
 				printf("\t\t\tI/O TX %u out (port %u): avg burst size = %.2f\n",
 					lcore,
-					(unsigned) port,
+					port,
 					((double) lp->tx.nic_ports_count[port]) / ((double) lp->tx.nic_ports_iters[port]));
 				lp->tx.nic_ports_iters[port] = 0;
 				lp->tx.nic_ports_count[port] = 0;
@@ -418,7 +416,7 @@ app_lcore_io_tx(
 static inline void
 app_lcore_io_tx_flush(struct app_lcore_params_io *lp)
 {
-	uint8_t port;
+	uint16_t port;
 	uint32_t i;
 
 	for (i = 0; i < lp->tx.n_nic_ports; i++) {
@@ -569,7 +567,7 @@ app_lcore_worker(
 			if (lp->rings_out_iters[port] == APP_STATS){
 				printf("\t\tWorker %u out (NIC port %u): enq success rate = %.2f\n",
 					(unsigned) lp->worker_id,
-					(unsigned) port,
+					port,
 					((double) lp->rings_out_count[port]) / ((double) lp->rings_out_iters[port]));
 				lp->rings_out_iters[port] = 0;
 				lp->rings_out_count[port] = 0;
